Corrected pv_correlation function
In the matlab version, when caclualting the pv part of the spatial
correlation, it uses imaginary numbers and then squares it. This
removes the sqrt(-1), but now makes the
pv negative. We do not have to handle that here, so instead of adding
a negative number, we subtract a positive one.

I'm unsure why the matlab version behaves in such a way. Testing
Will reveal whether or not this will become an issue.