Fixed two problems:
1 - When removing elements from the remaining indices array,
there was an ocassional problem of trying to access an element
that was out of scope. To make the remaing index more robust,
instead of attempting to remove elements, we can check if they
exist in one of the already selected indices array, and if they
don't, add them to a new array.

2 - Fixing the above problem also fixed an issue of not finding
entirely unique points, but I need to write some further tests
to check this