Simplify interpolation
This commit changes the strategy for interpolating climate data.

Instead of finding only points to interpolate closest to the best
pressure value, we find them all and then filter them down later
to find the closest in pressure.

The reason for doing this is that we can greatly simplify and
speed up the process in this way by vectorising a lot of the
calculations.